Fix compiler warnings about enums and arithmetics (#634)
* Fix compiler warnings about enums and arithmetics

Force casting enums into integer/float values when used for arithmetic operations.

* compiler warnings about enums and arithmetics

+++ src/game/Entities/Player.cpp
@@ -5236,7 +5236,7 @@ float Player::GetRatingMultiplier(CombatRating cr) const
 
 float Player::GetRatingBonusValue(CombatRating cr) const
 {
-    return float(GetUInt32Value(PLAYER_FIELD_COMBAT_RATING_1 + cr)) * GetRatingMultiplier(cr);
+    return float(GetUInt32Value(static_cast<uint16>(PLAYER_FIELD_COMBAT_RATING_1) + static_cast<uint16>(cr))) * GetRatingMultiplier(cr);
 }
 
 void Player::ApplyRatingMod(CombatRating cr, int32 value, bool apply)
@@ -5277,7 +5277,7 @@ void Player::UpdateRating(CombatRating cr)
     int32 amount = m_baseRatingValue[cr];
     if (amount < 0)
         amount = 0;
-    SetUInt32Value(PLAYER_FIELD_COMBAT_RATING_1 + cr, uint32(amount));
+    SetUInt32Value(static_cast<uint16>(PLAYER_FIELD_COMBAT_RATING_1) + static_cast<uint16>(cr), uint32(amount));
 
     bool affectStats = CanModifyStats();
